gona have to do daynemic quntization

+++ translate.py
@@ -1,4 +1,5 @@
 #from translate import get_quantmodel_and_tokenizer
+import intel_extension_for_pytorch as ipex
 from transformers import AutoModelForSeq2SeqLM, NllbTokenizerFast
 import torch
 import os 
@@ -103,11 +104,10 @@ def translate_text(text,spliter,tokenizer,model,num_beams=10,max_new_tokens=10**
         ans+='\n\n'
     return ans[:-2]
 
-import intel_extension_for_pytorch as ipex
 if __name__=="__main__":
 
     model,tokenizer=get_model_and_tokenizer()
-    model=model.to('xpu')
+    model=model.to(torch.bfloat16).to('xpu')
     spliter=stanza.Pipeline(lang='en',verbose=False)
 
     with open('test_text.txt') as f:
